+++ src/components/Sounds/SoundCard/styled/CardMediaStyled.js
@@ -3,7 +3,8 @@ import { styled } from '@material-ui/core/styles';
 
 const CardMediaStyled = styled(CardMedia)({
   height: 0,
-  paddingTop: '56.25%' //16:9
+  // paddingTop: '56.25%' //16:9
+  paddingTop: '75%' //1:1?
 });
 
 export default CardMediaStyled;