+++ src/components/TransportBar/components/TransportBar.js
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import Grid from '@material-ui/core/Grid';
 import { KeyboardArrowUp, KeyboardArrowDown } from '@material-ui/icons';
 
+import { ASMRButton } from '../../Button';
 import SoundEngine from '../../../sound-engine';
 import { RowBox } from '../../Box';
 import { SecondaryButton } from '../../Button';
@@ -10,9 +11,21 @@ import TransportBarWrapper from '../styled/TransportBarWrapper';
 import TransportControls from './TransportControls';
 import SoundInfo from './SoundInfo';
 import VolumeSlider from './VolumeSlider';
+import ASMRButtonContainer from '../styled/ASMRButtonContainer';
 
 export default function TransportBar(props) {
-  const { soundInfo, playlistShow, onPrev, onPlay, onNext, onStop, onPlaylistChange, disabled } = props;
+  const {
+    soundInfo,
+    playlistShow,
+    asmrActive,
+    onPrev,
+    onPlay,
+    onNext,
+    onStop,
+    onPlaylistChange,
+    onASMRClick,
+    disabled } = props;
+
   const { sounds, currentSound } = soundInfo;
   const result = sounds.find(item => item.id === currentSound);
   const [masterVolume, setMasterVolume] = useState(1.0);
@@ -24,7 +37,7 @@ export default function TransportBar(props) {
 
   return (
     <TransportBarWrapper container>
-      <Grid item md={2}>
+      <Grid item md={3}>
         <RowBox>
           <SecondaryButton
             tooltip={playlistShow ? 'Hide playlist' : 'Show playlist'}
@@ -38,7 +51,10 @@ export default function TransportBar(props) {
           <SoundInfo cover={result?.cover} title={result?.name} />
         </RowBox>
       </Grid>
-      <Grid item md={8}>
+      <Grid item md={6} style={{ position: 'relative' }}>
+        <ASMRButtonContainer>
+          <ASMRButton active={asmrActive} onClick={() => onASMRClick(!asmrActive)} />
+        </ASMRButtonContainer>
         <TransportControls
           onPlay={() => onPlay()}
           onPrev={() => onPrev()}
@@ -47,10 +63,11 @@ export default function TransportBar(props) {
           disabled={disabled}
         />
       </Grid>
-      <Grid item md={2}>
+      <Grid item md={3} style={{ display: 'flex', justifyContent: 'flex-end' }}>
         <VolumeSlider
           value={masterVolume}
           onChange={handleVolumeChange}
+          style={{ width: '75%' }}
         />
       </Grid>
     </TransportBarWrapper>