+++ src/components/TransportBar/styled/ASMRButtonContainer.js
@@ -5,7 +5,7 @@ const ASMRButtonContainer = styled(Box)({
   position: 'absolute',
   left: 0,
   right: 0,
-  bottom: 0,
+  bottom: '150%',
   display: 'flex',
   justifyContent: 'center'
 });