+++ CHANGES.md
@@ -7,3 +7,9 @@
 2. Fix: fields of containers randomize on each play event in loop.
 3. Fix: setting of non audio param field in sound objects is working.
 4. Fix: increased transition between new random values from 0 to 250ms.
+
+## Main UI
+5. Feat: Add cover on the left side of the sound name in transport bar.
+6. Fix: sound name displaying in transport bar.
+7. Feat: Change sizes of cards.
+8. Fix: ASMR button follows play button in vertical direction.